requirements.txt
pyyaml
uuid
requests